testStartCombine

+++ Assets/Script/BattleChar.cs
@@ -12,7 +12,8 @@ public class BattleChar : MonoBehaviour
     public List<BattleMove> movesAvailable;
 
     public string charName;
-    public int level;
+    //public int level;
+    public float level;
     public int currentHP;
     public int maxHP;
     public int currentMP;
